refactor(ui): loading dots 공통 패키지로 이동
