+++ apps/blog/src/pages/post/posts-page.tsx
@@ -4,6 +4,7 @@ import { getPosts } from "@/service/posts";
 import { sortByDate, extractDateFromMeta, formatDate } from "@/utils/date";
 import { AnalyticsPageData, useAllSiteAnalytics } from "@mfe/shared";
 import { Item, PostMeta } from "@/types/contents/post";
+import { LoadingDots } from "@srf/ui";
 
 export default function PostsPage() {
   const posts = getPosts();
@@ -17,8 +18,6 @@ export default function PostsPage() {
   const { loading, error, totals, pageDataList } = useAllSiteAnalytics({
     apiUrl: import.meta.env.VITE_GA_API_URL,
   });
-
-  console.log(totals, pageDataList, posts);
 
   return (
     <div className="max-w-2xl">
@@ -59,8 +58,6 @@ const PostItem = ({
           ?.views ?? 0);
   }, [pageDataList, post.slug, isCountLoading]);
 
-  console.log(count);
-
   return (
     <article
       key={post.slug}
@@ -83,15 +80,7 @@ const PostItem = ({
         )}
 
         <div className="text-xs text-gray-500 dark:text-gray-500 flex items-center gap-1">
-          {isCountLoading ? (
-            <span className="inline-flex items-center gap-1">
-              <span className="inline-block h-1.5 w-1.5 animate-pulse rounded-full bg-gray-400 dark:bg-gray-500" />
-              <span className="inline-block h-1.5 w-1.5 animate-pulse rounded-full bg-gray-400 dark:bg-gray-500 [animation-delay:0.2s]" />
-              <span className="inline-block h-1.5 w-1.5 animate-pulse rounded-full bg-gray-400 dark:bg-gray-500 [animation-delay:0.4s]" />
-            </span>
-          ) : (
-            <span>{count}</span>
-          )}
+          {isCountLoading ? <LoadingDots /> : <span>{count}</span>}
           <span>views</span>
         </div>
       </div>
