+++ packages/ui/src/components/loading/index.ts
@@ -1,0 +1,1 @@
+export { default as LoadingDots } from "./loading-dots";
